requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

ccxt >= 4.0.68
numpy >= 1.24.4
pandas >= 2.0.3
pyfiglet >= 0.8.post1
python_dotenv >= 1.0.0
statsmodels >= 0.14.0
